requirements.txt
certifi==2017.11.5
chardet==3.0.4
click==6.7
decorator==4.1.2
Django==1.11.5
django-bootstrap4==0.0.4
django-debug-toolbar==1.8
django-jquery==3.1.0
django-map-widgets==0.1.9
django-multiselectfield==0.1.8
flake8==3.5.0
GDAL==2.2.2
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.0
mccabe==0.6.1
mysqlclient==1.3.12
numpy==1.13.3
olefile==0.44
parso==0.1.0
pew==1.1.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.3.0
pip-chill==0.1.6
pipenv==9.0.0
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
ptyprocess==0.5.2
pycodestyle==2.3.1
pycountry==17.9.23
pyflakes==1.6.0
Pygments==2.2.0
pytz==2017.2
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.2.4
traitlets==4.3.2
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
wcwidth==0.1.7
